requirements.txt
streamlit==1.35.0
pandas==2.2.2
yfinance==0.2.40
backtrader==1.9.78.123
matplotlib==3.9.0
prophet==1.1.5
python-dateutil==2.9.0.post0
requests==2.32.3
torch==2.3.0
numpy==1.26.4
scikit-learn==1.5.0
scipy==1.13.1
torchvision==0.18.0
tqdm==4.66.4
Pillow==10.3.0
plotly==5.22.0
streamlit_tags==1.2.8
